# swagger/output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My-home API",
    "description": "API for my portfolio page",
    "version": "1.0.0"
  },
  "host": "localhost:8080/api",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "accessToken": {
      "type": "apiKey",
      "in": "cookie",
      "name": "accessToken",
      "description": "Need to get access"
    },
    "refreshToken": {
      "type": "apiKey",
      "in": "cookie",
      "name": "refreshToken",
      "description": "Need to refresh tokens"
    }
  },
  "paths": {
    "/admin/auth/login": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "authenticate",
        "description": "Admin login",
        "operationId": "authenticateUsingPOST",
        "parameters": [
          {
            "name": "login, password",
            "in": "body",
            "description": "Authorization data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AdminAuth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result info and cookies",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/auth/register": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "authenticate",
        "description": "Allows the administrator to create a new account",
        "operationId": "authenticateUsingPOST_1",
        "parameters": [
          {
            "name": "login, password",
            "in": "body",
            "description": "Authorization data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AdminAuth"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Result info",
            "in": "cookie",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/auth/refresh": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "authorization",
        "description": "Admin refresh",
        "operationId": "authenticateUsingPOST_2",
        "parameters": [
          {
            "name": "refreshToken",
            "in": "body",
            "description": "Refresh data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "refreshToken": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result info and cookies",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/admin/feedback": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "create feedback",
        "description": "Allows the administrator to create a new feedback",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "description": "New feedback data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "type": "string",
                  "example": "Test"
                },
                "lastName": {
                  "type": "string",
                  "example": "Person"
                },
                "text": {
                  "type": "string",
                  "example": "test text"
                },
                "rating": {
                  "type": "number",
                  "example": 5
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Result info",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/feedback/:id": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "edit feedback",
        "description": "Allows the administrator to edit a feedback",
        "parameters": [
          {
            "name": "id",
            "description": "Deleted project id",
            "required": true,
            "in": "query",
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "description": "Edited feedback data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Feedback"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result info",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/feedbacks/:id": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "delete feedback",
        "description": "Allows the administrator to delete a feedback",
        "parameters": [
          {
            "name": "id",
            "description": "Deleted feedback id",
            "required": true,
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Result info",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AdminAuth": {
      "type": "object",
      "properties": {
        "login": {
          "type": "string",
          "example": "admin"
        },
        "password": {
          "type": "string",
          "example": "admin"
        }
      }
    },
...
    "Feedback": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "Test"
        },
        "lastName": {
          "type": "string",
          "example": "Person"
        },
        "text": {
          "type": "string",
          "example": "test text"
        },
        "rating": {
          "type": "number",
          "example": 5
        }
      }
    },
    "Response": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "string"
        },
        "statusCod": {
          "type": "string",
          "example": "number"
        },
        "message": {
          "type": "string",
          "example": "string"
        }
      }
    }
  }
}